Support different vector length for iir vectorization.
