+++ midend/lib/Conversion/DAPVectorization/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_mlir_library(DAPVectorization
+  DAPVectorization.cpp
+
+  LINK_LIBS PUBLIC
+  BuddyDAPUtils
+)
