+++ midend/lib/Conversion/LowerDAP/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_mlir_library(LowerDAPPass
   LowerDAPPass.cpp
-  DAPVectorization.cpp
   )
