+++ tools/buddy-opt/CMakeLists.txt
@@ -18,6 +18,7 @@ target_link_libraries(buddy-opt
   LowerDIPPass
   BuddyDAP
   LowerDAPPass
+  DAPVectorization
   BuddyRVV
   LowerRVVPass
   MatMulOptimization
